requirements.txt
faiss==1.5.3
numpy==1.18.5
pandas==1.0.4
PyYAML==5.3.1
scikit-learn==0.23.1
torch==1.6.0
tqdm==4.46.1
transformers==4.3.2
